+++ src/app/services/coordinator/coordinator.service.ts
@@ -9,6 +9,7 @@ import { ErrorHandlingPipe } from '../../pipes/error-handling.pipe';
 import { Account } from '../wallet/wallet';
 import Big from 'big.js';
 import { TokenService } from '../token/token.service';
+import { CONSTANTS } from '../../../environments/environment';
 
 export interface ScheduleData {
   pkh: string;
@@ -25,10 +26,10 @@ enum State {
 @Injectable()
 export class CoordinatorService {
   scheduler: Map<string, any> = new Map<string, any>(); // pkh + delay
-  defaultDelayActivity = 30000; // 30s
+  defaultDelayActivity = CONSTANTS.MAINNET ? 60000 : 30000; // 60/30s
   shortDelayActivity = 5000; // 5s
   tzrateInterval: any;
-  defaultDelayPrice = 300000; // 300s
+  defaultDelayPrice = CONSTANTS.MAINNET ? 300000 : 3600000; // 5/60m
   accounts: Account[];
   constructor(
     private activityService: ActivityService,
@@ -96,7 +97,7 @@ export class CoordinatorService {
         setTimeout(() => {
           // Failsafe
           if (
-            this.scheduler &&
+            this.scheduler?.size &&
             this.scheduler.get(pkh).stateCounter === counter
           ) {
             console.log('Timeout from wait state');
@@ -153,7 +154,7 @@ export class CoordinatorService {
       err => console.log('Error in update(): ' + JSON.stringify(err)),
       () => {
         console.log(`account[${this.accounts.findIndex((a) => a.address === pkh)}][${
-          typeof this.scheduler.get(pkh).state !== 'undefined' ? this.scheduler.get(pkh).state : '*'}]: <<`);
+          typeof this.scheduler.get(pkh)?.state !== 'undefined' ? this.scheduler.get(pkh).state : '*'}]: <<`);
       }
     );
   }
@@ -235,7 +236,8 @@ export class CoordinatorService {
           hash: metadata.opHash,
           block: null,
           timestamp: new Date().getTime(),
-          tokenId: metadata.tokenTransfer ? metadata.tokenTransfer : undefined
+          tokenId: metadata.tokenTransfer ? metadata.tokenTransfer : undefined,
+          entrypoint: op.parameters?.entrypoint ? op.parameters.entrypoint : ''
         };
         let account = this.walletService.wallet.getAccount(from);
         account.activities.unshift(transaction);