+++ src/app/services/wallet/wallet.service.ts
@@ -295,7 +295,6 @@ export class WalletService {
       } else if (this.wallet instanceof TorusWallet) {
         type = 'TorusWallet';
       }
-      console.log('Type is ' + type);
       localStorage.setItem(
         this.storeKey,
         JSON.stringify({ type, localStorageId: this.storageId, data: this.wallet })
@@ -372,7 +371,6 @@ export class WalletService {
     this.wallet.totalBalanceUSD = wd.totalBalanceUSD;
     this.wallet.totalBalanceXTZ = wd.totalBalanceXTZ;
     if (wd.lookups) {
-      console.log('found', wd.lookups);
       this.wallet.lookups = wd.lookups;
     }
     for (const implicit of wd.implicitAccounts) {
@@ -403,9 +401,9 @@ export class WalletService {
         origAcc.balanceXTZ = originated.balanceXTZ;
         origAcc.delegate = originated.delegate;
         if (originated.activitiesCounter) { // prevent storage from breaking (1.11)
-          impAcc.state = originated.activitiesCounter.toString();
+          origAcc.state = originated.activitiesCounter.toString();
         } else {
-          impAcc.state = originated.state;
+          origAcc.state = originated.state;
         }
         origAcc.activities = this.activityMigration(originated.activities);
         impAcc.originatedAccounts.push(origAcc);
@@ -415,10 +413,10 @@ export class WalletService {
   }
   activityMigration(activities: any[]): Activity[] { // prevent storage from breaking (1.11)
     return activities.map(activity => {
-      if (!activity.source.address) {
+      if (activity.source.address === undefined) {
         activity.source = { address: activity.source };
       }
-      if (!activity.destination.address) {
+      if (activity.destination.address === undefined) {
         activity.destination = { address: activity.destination };
       }
       return activity;